feat: Add donation feature to header
Introduces a donation button to the application header and a corresponding modal. This allows users to contribute to the developer's efforts.

The modal is triggered by clicking the "Donate Now" button and provides information on how to donate, currently via Jazzcash. State management for the modal's visibility has been added to the App component.